Adding 17G support for idrac LC logs
